changed so that every go file is in the same main package
