+++ .gitignore
@@ -12,4 +12,4 @@
 *.out
 
 # MAC
-DS_Store
+.DS_Store
