+++ d7024e/bootstrap.sh
@@ -1,0 +1,4 @@
+#! /bin/sh
+/usr/local/go/bin/go get -u github.com/golang/protobuf/{proto,protoc-gen-go}
+/usr/local/go/bin/go build
+./d7024e
