+++ d7024e/kademlia.go
@@ -1,4 +1,4 @@
-package d7024e
+package main
 
 type Kademlia struct {
 	routingTB RoutingTable
