+++ d7024e/node.sh
@@ -1,0 +1,3 @@
+go get -u github.com/golang/protobuf/{proto,protoc-gen-go}
+go build
+./d7024e -m client
