--- docker-compose-lab.yml
@@ -1,21 +1,0 @@
-version: "3"
-services:
-  kademliaNodes:
-    image: larjim/kademlialab:latest
-    stdin_open: true
-    tty: true
-    deploy:
-      replicas: 5  
-#        resources:
-#           limits:
-#              cpus: "0.1"
-#              memory: 50M
-      restart_policy:
-        condition: on-failure
-#    ports:
-#      - "4000:80"
-    networks:
-      - kademlia_network
-      
-networks:
-  kademlia_network:
